# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Дом в порядке API
  version: 0.1.0
  description: Собственный API MVP. Все маршруты `/api/*`, кроме health, требуют cookie-сессию диспетчера.
servers:
  - url: https://jorwey.ru
    description: Production
paths:
...
  /api/buildings:
    get: { security: [{ cookieAuth: [] }], summary: Список домов, responses: { '200': { description: Список домов } } }
    post:
      security: [{ cookieAuth: [] }]
      summary: Добавить дом и квартиры
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BuildingInput' } } } }
      responses: { '201': { description: Дом создан }, '400': { description: Ошибка DTO } }
  /api/buildings/{id}:
    patch:
      security: [{ cookieAuth: [] }]
      summary: Изменить дом
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/BuildingInput' } } } }
      responses: { '200': { description: Дом изменён }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      security: [{ cookieAuth: [] }]
      summary: Удалить дом и связанные демонстрационные/операционные данные
      description: Требуется роль supervisor. Привязки жителей удаляются, поэтому им потребуется повторная регистрация.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: Дом удалён
          content:
            application/json:
              schema:
                type: object
                required: [revokedMemberships]
                properties: { revokedMemberships: { type: integer, minimum: 0 } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/emergency-services:
    get: { security: [{ cookieAuth: [] }], summary: Справочник служб, responses: { '200': { description: Список } } }
    post: { security: [{ cookieAuth: [] }], summary: Добавить контакт, responses: { '201': { description: Контакт создан } } }
  /api/emergency-services/{id}:
    patch:
      security: [{ cookieAuth: [] }]
      summary: Изменить контакт службы
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Контакт сохранён }, '400': { description: Ошибка DTO } }
    delete:
      security: [{ cookieAuth: [] }]
      summary: Удалить контакт службы
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '204': { description: Контакт удалён } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: smart_home_admin }
...
  responses:
...
    Forbidden: { description: Недостаточно прав, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Не найдено, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
...
    Error: { type: object, required: [code], properties: { code: { type: string }, requestId: { type: string } } }
...
    BuildingInput: { type: object, required: [address, apartmentsCount, entrancesCount], properties: { address: { type: string }, apartmentsCount: { type: integer, minimum: 1, maximum: 2000 }, entrancesCount: { type: integer, minimum: 1, maximum: 50 }, chatLink: { type: [string, 'null'], format: uri } } }
...